feat: add delete user route

+++ internal/services/folder_service/folder_service.go
@@ -35,7 +35,7 @@ func (f *FolderService) FolderExist(fldName string) bool {
 		return false
 	}
 
-	if fldModel == (folder_model.FolderModel{}) {
+	if fldModel == nil {
 		return false
 	}
 
@@ -52,11 +52,14 @@ func (f *FolderService) CreateFolder(fldName string, usrID uuid.UUID) error {
 		return errsvc.ErrGenFolderFailed
 	}
 
+	fldId := uuid.New()
+
 	fldModel, err := f.FldRepo.CreateFld(folder_model.FolderModel{
-		ID:        uuid.New(),
+		ID:        fldId,
 		Name:      fldName,
 		ParentID:  nil,
 		OwnerID:   usrID,
+		MainFldId: nil,
 		CreatedAt: time.Now(),
 		UpdatedAt: time.Now(),
 	})
@@ -86,12 +89,20 @@ func (f *FolderService) CreateFolder(fldName string, usrID uuid.UUID) error {
 	return nil
 }
 
-func (f *FolderService) DelMainFld(fldName string) error {
-	fldModel, err := f.FldRepo.GetGeneralFolderByName(fldName)
+func (f *FolderService) DelFld(fldName string) error {
+	fldModel, err := f.FldRepo.GetFldByName(fldName)
 
 	if err != nil {
 		log.Error().Err(err).Msg("failed to get folder")
 		return errsvc.ErrFldNotFound
+	}
+
+	if fldModel == nil {
+		return errsvc.ErrFldNotFound
+	}
+
+	if fldModel.MainFldId == nil {
+		return errsvc.ErrCantDelMainFld
 	}
 
 	err = os.RemoveAll(filepath.Join(f.StoragePath, fldName))
@@ -101,7 +112,7 @@ func (f *FolderService) DelMainFld(fldName string) error {
 		return errsvc.ErrFldDeleteFailed
 	}
 
-	err = f.FldRepo.DelMainFld(fldModel.ID)
+	err = f.FldRepo.DelFld(fldModel.ID)
 
 	if err != nil {
 		log.Error().Err(err).Msg("failed to delete folder")
